Time: 57 ms (11.91%), Space: 23.8 MB (95.67%) - LeetHub

+++ 238-product-of-array-except-self/238-product-of-array-except-self.cpp
@@ -1,14 +1,13 @@
 class Solution {
 public:
-    vector<int> productExceptSelf(vector<int>& nums) {
-        int n=(int)nums.size();
-        int fromBegin=1,fromLast=1;
-        vector<int> res(n,1);
-        
-        for(int i=0;i<n;i++){
+    vector<int> productExceptSelf(vector<int> &nums) {
+        int n = nums.size();
+        int fromBegin = 1, fromLast = 1;
+        vector<int> res(n, 1);
+        for (int i = 0; i < n; i++){
             res[i] *= fromBegin;
             fromBegin *= nums[i];
-            res[n-1-i] *= fromLast;
+            res[n - 1 - i] *= fromLast;
             fromLast *= nums[n-1-i];
         }
         return res;
